magenta/requirements.txt
acres==0.2.0
annotated-types==0.7.0
anyio==4.7.0
bcrypt==4.2.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
ci-info==0.3.0
click==8.1.7
configobj==5.0.9
configparser==7.1.0
cryptography==44.0.0
distro==1.9.0
dnspython==2.7.0
etelemetry==0.3.1
exceptiongroup==1.2.2
fastapi==0.115.6
filelock==3.16.1
fpdf==1.7.2
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
importlib_resources==6.4.5
iniconfig==2.0.0
isodate==0.6.1
jiter==0.8.2
loguru==0.7.3
looseversion==1.3.0
lxml==5.3.0
networkx==3.4.2
nibabel==5.3.2
nipype==1.9.1
numpy==2.2.0
openai==1.57.4
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pathlib==1.0.1
pgvector==0.3.6
pluggy==1.5.0
prov==2.0.1
psycopg2-binary==2.9.10
puremagic==1.28
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
pydot==3.0.3
PyJWT==2.10.1
pymongo==4.10.1
PyMuPDF==1.25.1
pyparsing==3.2.0
pytest==8.3.4
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2024.2
pyxnat==1.6.2
rdflib==6.3.2
requests==2.32.3
scipy==1.14.1
simplejson==3.19.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
tomli==2.2.1
tqdm==4.67.1
traits==6.4.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.34.0
